reorganize to handle metro symlink fail
The react native bundler, 'metro', doesn't follow symlinks. This meant
that in the example app we couldn't do

import Foo from "react-native-feed-media-audio-player"

when the library was installed via 'npm install ../package', because
that npm a symlink in
'node_modules/react-native-feed-media-audio-player' and metro wouldn't
follow it. Maddening. Hopefully this will be fixed with:

https://github.com/facebook/metro/pull/257

This patch inverts things, so the top level 'package' directory is
a symlink, and the node_modules/react... is the actual source. Now
we can run the example app while changing code in the library.